requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
pydantic==1.8.2
librosa==0.8.1
tensorflow==2.6.0
paho-mqtt==1.5.1
influxdb-client==1.21.0
python-dotenv==0.19.0
